Create hex_color_codes.py
The majority will want to write their own code to create a color code resource, but it seemed incomplete without adding the means to make use of the data file. This may not be the most elegant code, but it will return a dictionary, as intended. Which dictionary to use will depends on which codes are needed -  hexadecimal format or the old RGB three number tuple. Hex usually works in most instances.