Create color_codes.py
The majority will prefer to write their own code to utilize the data file though it seemed incomplete without providing the means with the file itself. I'll never claim it to be the most elegant code, but it functions as intended -  a color code resource primarily used in Python code.